change p_drop_hidden and p_drop_input -> p_keep_
The p_drop_ prefix was used in the model() function, but p_keep_ was used in the initialization and while training the model. This could lead to confusion – the [docs](http://tensorflow.org/api_docs/python/nn.md#dropout) describe the variable as keep_prob, so I changed the names to keep for consistency.